minio/proof-of-concept/requirements.txt
boto3
botocore
requests
python-dotenv